Refactor knowledge graph pipeline into knowledge package
